# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Decision Brain API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /outcomes:
    post:
      operationId: createOutcome
      tags: [outcomes]
      summary: Log an outcome for a decision
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOutcomeBody"
      responses:
        "201":
          description: Outcome created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outcome"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      security:
        - bearerAuth: []
      parameters:
        - name: unread
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/read:
    patch:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark alert as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Dashboard analytics overview
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
  /analytics/outcomes:
    get:
      operationId: getOutcomeAnalytics
      tags: [analytics]
      summary: Outcome performance analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Outcome analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutcomeAnalytics"
  /analytics/patterns:
    get:
      operationId: getPatternAnalytics
      tags: [analytics]
      summary: Decision pattern analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pattern analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatternAnalytics"
  /analytics/blindspots:
    get:
      operationId: getBlindspotAnalytics
      tags: [analytics]
      summary: Blind spot analysis
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Blind spot analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlindspotAnalytics"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Decision:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        sourcePlatform:
          type: string
          nullable: true
          enum: [gmail, zoom, slack, meet, teams, notion, manual, null]
        stakes:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [active, reversed, pending]
        tags:
          type: array
          items:
            type: string
        decidedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        outcomeScore:
          type: integer
          nullable: true
        alertCount:
          type: integer
      required:
        - id
        - userId
        - title
        - stakes
        - status
        - tags
        - createdAt
        - alertCount
...
    Outcome:
      type: object
      properties:
        id:
          type: string
        decisionId:
          type: string
        userId:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 100
        notes:
          type: string
          nullable: true
        checkInterval:
          type: string
          enum: ["30d", "90d", "180d", "1y"]
        trackedAt:
          type: string
          format: date-time
      required:
        - id
        - decisionId
        - userId
        - score
        - checkInterval
        - trackedAt
...
    CreateOutcomeBody:
      type: object
      properties:
        decisionId:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 100
        notes:
          type: string
        checkInterval:
          type: string
          enum: ["30d", "90d", "180d", "1y"]
          default: "90d"
      required:
        - decisionId
        - score
...
    Alert:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        decisionId:
          type: string
          nullable: true
        alertType:
          type: string
          enum: [pattern_match, blind_spot, repeat_mistake, similarity]
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - alertType
        - title
        - description
        - severity
        - isRead
        - createdAt
...
    AnalyticsOverview:
      type: object
      properties:
        totalDecisions:
          type: integer
        decisionsThisMonth:
          type: integer
        decisionsThisWeek:
          type: integer
        avgOutcomeScore:
          type: number
        unreadAlerts:
          type: integer
        pendingOutcomes:
          type: integer
        decisionsByStakes:
          type: object
          additionalProperties:
            type: integer
        decisionsByPlatform:
          type: object
          additionalProperties:
            type: integer
        recentDecisions:
          type: array
          items:
            $ref: "#/components/schemas/Decision"
        outcomeScoreTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        decisionVelocity:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
      required:
        - totalDecisions
        - decisionsThisMonth
        - decisionsThisWeek
        - avgOutcomeScore
        - unreadAlerts
        - pendingOutcomes
        - decisionsByStakes
        - decisionsByPlatform
        - recentDecisions
        - outcomeScoreTrend
        - decisionVelocity
    OutcomeAnalytics:
      type: object
      properties:
        winRate:
          type: number
        avgScore:
          type: number
        scoreByStakes:
          type: object
          additionalProperties:
            type: number
        scoreByPlatform:
          type: object
          additionalProperties:
            type: number
        scoreTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        topDecisions:
          type: array
          items:
            $ref: "#/components/schemas/Decision"
      required:
        - winRate
        - avgScore
        - scoreByStakes
        - scoreByPlatform
        - scoreTrend
        - topDecisions
    PatternAnalytics:
      type: object
      properties:
        totalPatterns:
          type: integer
        patternsByType:
          type: object
          additionalProperties:
            type: integer
        repeatMistakes:
          type: integer
        blindSpots:
          type: integer
        recentPatterns:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
      required:
        - totalPatterns
        - patternsByType
        - repeatMistakes
        - blindSpots
        - recentPatterns
    BlindspotAnalytics:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/BlindspotCategory"
        urgencyBiasCount:
          type: integer
        noExternalDataCount:
          type: integer
        repeatedFailureCount:
          type: integer
      required:
        - categories
        - urgencyBiasCount
        - noExternalDataCount
        - repeatedFailureCount
    BlindspotCategory:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
        avgOutcomeScore:
          type: number
        riskLevel:
          type: string
          enum: [low, medium, high]
      required:
        - category
        - count
        - avgOutcomeScore
        - riskLevel
    TrendPoint:
      type: object
      properties:
        date:
          type: string
        value:
          type: number
      required:
        - date
        - value
...